Refactored `Util`
