--- CLion/Air Analyzer/lib/ObserverPattern/src/AbstractObserver.h
@@ -1,10 +1,0 @@
-#ifndef ABSTRACTOBSERVER_H
-    #define ABSTRACTOBSERVER_H
-
-    #include <Arduino.h>
-
-    class AbstractObserver {
-        public:
-            virtual void update() = 0;
-    };
-#endif
