+++ CLion/Air Analyzer/lib/Util/src/AbstractObserver.h
@@ -1,0 +1,10 @@
+#ifndef ABSTRACTOBSERVER_H
+#define ABSTRACTOBSERVER_H
+
+#include <Arduino.h>
+
+class AbstractObserver {
+    public:
+        virtual void update() = 0;
+};
+#endif
